Update tests for models and typeof
