+++ numba_dppy/tests/njit_tests/dpnp_ndarray/test_boxing.py
@@ -12,40 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for dpnp.ndarray interoperability with Numba
-
-See numba/tests/test_ndarray_subclasses.py
 """
-
+Tests for boxing for dpnp.ndarray
+"""
 
 import pytest
 from dpctl.tensor.numpy_usm_shared import ndarray as dpctl_ndarray
 from dpnp import ndarray as dpnp_ndarray
-from numba import njit, typeof, types
-
-from numba_dppy.dpnp_ndarray import dpnp_ndarray_Type
-from numba_dppy.numpy_usm_shared import UsmSharedArrayType
-
-
-@pytest.mark.parametrize(
-    "array_type, expected_numba_type",
-    [
-        (dpctl_ndarray, UsmSharedArrayType),
-        (dpnp_ndarray, dpnp_ndarray_Type),
-    ],
-)
-@pytest.mark.parametrize(
-    "shape, expected_ndim",
-    [
-        ([1], 1),
-        ([1, 1], 2),
-    ],
-)
-def test_typeof(array_type, shape, expected_numba_type, expected_ndim):
-    array = array_type(shape)
-    expected_type = expected_numba_type(types.float64, expected_ndim, "C")
-    assert typeof(array) == expected_type
-
+from numba import njit
 
 dpnp_mark = pytest.mark.xfail(raises=TypeError, reason="No unboxing")
